Conflicts:
	config/color_themes.yml
	config/eye.rb
	db/schema.rb

+++ config/eye.rb
@@ -12,13 +12,23 @@ Eye.application("diaspora") do
   stderr "log/eye_processes_stderr.log"
 
   process :web do
-    start_command "bin/bundle exec unicorn -c config/unicorn.rb -D"
-    daemonize false
+    unicorn_command = "bin/bundle exec unicorn -c config/unicorn.rb"
+
+    if rails_env == "production"
+      start_command "#{unicorn_command} -D"
+      daemonize false
+      restart_command "kill -USR2 {PID}"
+      restart_grace 10.seconds
+    else
+      start_command unicorn_command
+      daemonize true
+    end
+
     pid_file AppConfig.server.pid.get
     stop_signals [:TERM, 10.seconds]
-    restart_command "kill -USR2 {PID}"
 
     env "PORT" => ENV["PORT"]
+
     monitor_children do
       stop_command "kill -QUIT {PID}"
     end
